scripts: update format of function chunks JSON output
After this commit, the format is as follows. Map from function address
to list of function chunks (of which the function is composed). Each
chunk is denoted by its start address and size in bytes.

Example with two function at address 0x0 and 0xB. Note that the function
at address 0x0 is discontinuous, as the chunk 0xB exist in between the
two chunks 0x0 and 0x11.

	{
		"0x00000000": [
			{
				"chunk_addr": "0x00000000",
				"chunk_size": 11
			},
			{
				"chunk_addr": "0x00000011",
				"chunk_size": 6
			}
		],
		"0x0000000B": [
			{
				"chunk_addr": "0x0000000B",
				"chunk_size": 6
			}
		]
	}

Updates mewmew/lift#11.